backend/requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
Markdown==3.5.2
PyJWT==1.7.1
pytz==2024.1
sqlparse==0.4.4
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
dataclasses-json==0.6.7
distro==1.9.0
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
groq==0.11.0
grpcio==1.66.1
grpcio-tools==1.66.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
hyperframe==6.0.1
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.6
langchain-experimental==0.3.1.post1
langchain-groq==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.128
MarkupSafe==2.1.5
marshmallow==3.22.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pillow==10.4.0
portalocker==2.10.1
protobuf==5.28.2
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pypdf==5.0.0
python-dotenv==1.0.1
PyYAML==6.0.2
qdrant-client==1.11.3
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.1
setuptools==75.1.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.20.0
torch==2.4.1
tqdm==4.66.5
transformers==4.45.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.12.1
